# kratos-server/api/rulego/v1/log_http.pb.go
// Code generated by protoc-gen-go-http. DO NOT EDIT.
// versions:
// - protoc-gen-go-http v2.5.3
// - protoc             v3.21.9
// source: rulego/v1/log.proto

package v1

import (
	context "context"
	http "github.com/go-kratos/kratos/v2/transport/http"
	binding "github.com/go-kratos/kratos/v2/transport/http/binding"
)

// This is a compile-time assertion to ensure that this generated file
// is compatible with the kratos package it is being compiled against.
var _ = new(context.Context)
var _ = binding.EncodeURL

const _ = http.SupportPackageIsVersion1

const OperationLogServiceDeleteRunLogs = "/api.rulego.v1.LogService/DeleteRunLogs"
const OperationLogServiceGetDebugLogs = "/api.rulego.v1.LogService/GetDebugLogs"
const OperationLogServiceListRunLogs = "/api.rulego.v1.LogService/ListRunLogs"

type LogServiceHTTPServer interface {
	DeleteRunLogs(context.Context, *DeleteRunLogsRequest) (*DeleteRunLogsResponse, error)
	GetDebugLogs(context.Context, *GetDebugLogsRequest) (*GetDebugLogsResponse, error)
	ListRunLogs(context.Context, *ListRunLogsRequest) (*ListRunLogsResponse, error)
}

func RegisterLogServiceHTTPServer(s *http.Server, srv LogServiceHTTPServer) {
	r := s.Route("/")
	r.GET("/api/v1/logs/debug", _LogService_GetDebugLogs0_HTTP_Handler(srv))
	r.GET("/api/v1/logs/runs", _LogService_ListRunLogs0_HTTP_Handler(srv))
	r.DELETE("/api/v1/logs/runs", _LogService_DeleteRunLogs0_HTTP_Handler(srv))
}

func _LogService_GetDebugLogs0_HTTP_Handler(srv LogServiceHTTPServer) func(ctx http.Context) error {
	return func(ctx http.Context) error {
		var in GetDebugLogsRequest
		if err := ctx.BindQuery(&in); err != nil {
			return err
		}
		http.SetOperation(ctx, OperationLogServiceGetDebugLogs)
		h := ctx.Middleware(func(ctx context.Context, req interface{}) (interface{}, error) {
			return srv.GetDebugLogs(ctx, req.(*GetDebugLogsRequest))
		})
		out, err := h(ctx, &in)
		if err != nil {
			return err
		}
		reply := out.(*GetDebugLogsResponse)
		return ctx.Result(200, reply)
	}
}

func _LogService_ListRunLogs0_HTTP_Handler(srv LogServiceHTTPServer) func(ctx http.Context) error {
	return func(ctx http.Context) error {
		var in ListRunLogsRequest
		if err := ctx.BindQuery(&in); err != nil {
			return err
		}
		http.SetOperation(ctx, OperationLogServiceListRunLogs)
		h := ctx.Middleware(func(ctx context.Context, req interface{}) (interface{}, error) {
			return srv.ListRunLogs(ctx, req.(*ListRunLogsRequest))
		})
		out, err := h(ctx, &in)
		if err != nil {
			return err
		}
		reply := out.(*ListRunLogsResponse)
		return ctx.Result(200, reply)
	}
}

func _LogService_DeleteRunLogs0_HTTP_Handler(srv LogServiceHTTPServer) func(ctx http.Context) error {
	return func(ctx http.Context) error {
		var in DeleteRunLogsRequest
		if err := ctx.BindQuery(&in); err != nil {
			return err
		}
		http.SetOperation(ctx, OperationLogServiceDeleteRunLogs)
		h := ctx.Middleware(func(ctx context.Context, req interface{}) (interface{}, error) {
			return srv.DeleteRunLogs(ctx, req.(*DeleteRunLogsRequest))
		})
		out, err := h(ctx, &in)
		if err != nil {
			return err
		}
		reply := out.(*DeleteRunLogsResponse)
		return ctx.Result(200, reply)
	}
}

type LogServiceHTTPClient interface {
	DeleteRunLogs(ctx context.Context, req *DeleteRunLogsRequest, opts ...http.CallOption) (rsp *DeleteRunLogsResponse, err error)
	GetDebugLogs(ctx context.Context, req *GetDebugLogsRequest, opts ...http.CallOption) (rsp *GetDebugLogsResponse, err error)
	ListRunLogs(ctx context.Context, req *ListRunLogsRequest, opts ...http.CallOption) (rsp *ListRunLogsResponse, err error)
}

type LogServiceHTTPClientImpl struct {
	cc *http.Client
}

func NewLogServiceHTTPClient(client *http.Client) LogServiceHTTPClient {
	return &LogServiceHTTPClientImpl{client}
}

func (c *LogServiceHTTPClientImpl) DeleteRunLogs(ctx context.Context, in *DeleteRunLogsRequest, opts ...http.CallOption) (*DeleteRunLogsResponse, error) {
	var out DeleteRunLogsResponse
	pattern := "/api/v1/logs/runs"
	path := binding.EncodeURL(pattern, in, true)
	opts = append(opts, http.Operation(OperationLogServiceDeleteRunLogs))
	opts = append(opts, http.PathTemplate(pattern))
	err := c.cc.Invoke(ctx, "DELETE", path, nil, &out, opts...)
	if err != nil {
		return nil, err
	}
	return &out, err
}

func (c *LogServiceHTTPClientImpl) GetDebugLogs(ctx context.Context, in *GetDebugLogsRequest, opts ...http.CallOption) (*GetDebugLogsResponse, error) {
	var out GetDebugLogsResponse
	pattern := "/api/v1/logs/debug"
	path := binding.EncodeURL(pattern, in, true)
	opts = append(opts, http.Operation(OperationLogServiceGetDebugLogs))
	opts = append(opts, http.PathTemplate(pattern))
	err := c.cc.Invoke(ctx, "GET", path, nil, &out, opts...)
	if err != nil {
		return nil, err
	}
	return &out, err
}

func (c *LogServiceHTTPClientImpl) ListRunLogs(ctx context.Context, in *ListRunLogsRequest, opts ...http.CallOption) (*ListRunLogsResponse, error) {
	var out ListRunLogsResponse
	pattern := "/api/v1/logs/runs"
	path := binding.EncodeURL(pattern, in, true)
	opts = append(opts, http.Operation(OperationLogServiceListRunLogs))
	opts = append(opts, http.PathTemplate(pattern))
	err := c.cc.Invoke(ctx, "GET", path, nil, &out, opts...)
	if err != nil {
		return nil, err
	}
	return &out, err
}

const OperationLocaleServiceGetLocales = "/api.rulego.v1.LocaleService/GetLocales"
const OperationLocaleServiceSaveLocales = "/api.rulego.v1.LocaleService/SaveLocales"

type LocaleServiceHTTPServer interface {
	GetLocales(context.Context, *GetLocalesRequest) (*GetLocalesResponse, error)
	SaveLocales(context.Context, *SaveLocalesRequest) (*SaveLocalesResponse, error)
}

func RegisterLocaleServiceHTTPServer(s *http.Server, srv LocaleServiceHTTPServer) {
	r := s.Route("/")
	r.GET("/api/v1/locales", _LocaleService_GetLocales0_HTTP_Handler(srv))
	r.POST("/api/v1/locales", _LocaleService_SaveLocales0_HTTP_Handler(srv))
}

func _LocaleService_GetLocales0_HTTP_Handler(srv LocaleServiceHTTPServer) func(ctx http.Context) error {
	return func(ctx http.Context) error {
		var in GetLocalesRequest
		if err := ctx.BindQuery(&in); err != nil {
			return err
		}
		http.SetOperation(ctx, OperationLocaleServiceGetLocales)
		h := ctx.Middleware(func(ctx context.Context, req interface{}) (interface{}, error) {
			return srv.GetLocales(ctx, req.(*GetLocalesRequest))
		})
		out, err := h(ctx, &in)
		if err != nil {
			return err
		}
		reply := out.(*GetLocalesResponse)
		return ctx.Result(200, reply)
	}
}

func _LocaleService_SaveLocales0_HTTP_Handler(srv LocaleServiceHTTPServer) func(ctx http.Context) error {
	return func(ctx http.Context) error {
		var in SaveLocalesRequest
		if err := ctx.Bind(&in); err != nil {
			return err
		}
		http.SetOperation(ctx, OperationLocaleServiceSaveLocales)
		h := ctx.Middleware(func(ctx context.Context, req interface{}) (interface{}, error) {
			return srv.SaveLocales(ctx, req.(*SaveLocalesRequest))
		})
		out, err := h(ctx, &in)
		if err != nil {
			return err
		}
		reply := out.(*SaveLocalesResponse)
		return ctx.Result(200, reply)
	}
}

type LocaleServiceHTTPClient interface {
	GetLocales(ctx context.Context, req *GetLocalesRequest, opts ...http.CallOption) (rsp *GetLocalesResponse, err error)
	SaveLocales(ctx context.Context, req *SaveLocalesRequest, opts ...http.CallOption) (rsp *SaveLocalesResponse, err error)
}

type LocaleServiceHTTPClientImpl struct {
	cc *http.Client
}

func NewLocaleServiceHTTPClient(client *http.Client) LocaleServiceHTTPClient {
	return &LocaleServiceHTTPClientImpl{client}
}

func (c *LocaleServiceHTTPClientImpl) GetLocales(ctx context.Context, in *GetLocalesRequest, opts ...http.CallOption) (*GetLocalesResponse, error) {
	var out GetLocalesResponse
	pattern := "/api/v1/locales"
	path := binding.EncodeURL(pattern, in, true)
	opts = append(opts, http.Operation(OperationLocaleServiceGetLocales))
	opts = append(opts, http.PathTemplate(pattern))
	err := c.cc.Invoke(ctx, "GET", path, nil, &out, opts...)
	if err != nil {
		return nil, err
	}
	return &out, err
}

func (c *LocaleServiceHTTPClientImpl) SaveLocales(ctx context.Context, in *SaveLocalesRequest, opts ...http.CallOption) (*SaveLocalesResponse, error) {
	var out SaveLocalesResponse
	pattern := "/api/v1/locales"
	path := binding.EncodeURL(pattern, in, false)
	opts = append(opts, http.Operation(OperationLocaleServiceSaveLocales))
	opts = append(opts, http.PathTemplate(pattern))
	err := c.cc.Invoke(ctx, "POST", path, in, &out, opts...)
	if err != nil {
		return nil, err
	}
	return &out, err
}