# kratos-server/openapi.yaml
...
openapi: 3.0.3
info:
    title: ""
    version: 0.0.1
paths:
...
    /api/v1/locales:
        get:
            tags:
                - LocaleService
            description: 获取本地化配置
            operationId: LocaleService_GetLocales
            parameters:
                - name: lang
                  in: query
                  schema:
                    type: string
                - name: username
                  in: query
                  schema:
                    type: string
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/api.rulego.v1.GetLocalesResponse'
        post:
            tags:
                - LocaleService
            description: 保存本地化配置
            operationId: LocaleService_SaveLocales
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/api.rulego.v1.SaveLocalesRequest'
                required: true
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/api.rulego.v1.SaveLocalesResponse'
...
    /api/v1/logs/debug:
        get:
            tags:
                - LogService
            description: 获取调试日志
            operationId: LogService_GetDebugLogs
            parameters:
                - name: chainId
                  in: query
                  schema:
                    type: string
                - name: nodeId
                  in: query
                  schema:
                    type: string
                - name: username
                  in: query
                  schema:
                    type: string
                - name: page
                  in: query
                  schema:
                    type: integer
                    format: int32
                - name: pageSize
                  in: query
                  schema:
                    type: integer
                    format: int32
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/api.rulego.v1.GetDebugLogsResponse'
    /api/v1/logs/runs:
        get:
            tags:
                - LogService
            description: 获取运行日志列表
            operationId: LogService_ListRunLogs
            parameters:
                - name: username
                  in: query
                  schema:
                    type: string
                - name: page
                  in: query
                  schema:
                    type: integer
                    format: int32
                - name: pageSize
                  in: query
                  schema:
                    type: integer
                    format: int32
                - name: chainId
                  in: query
                  schema:
                    type: string
                - name: startTime
                  in: query
                  schema:
                    type: string
                - name: endTime
                  in: query
                  schema:
                    type: string
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/api.rulego.v1.ListRunLogsResponse'
        delete:
            tags:
                - LogService
            description: 删除运行日志
            operationId: LogService_DeleteRunLogs
            parameters:
                - name: ids
                  in: query
                  schema:
                    type: array
                    items:
                        type: string
                - name: username
                  in: query
                  schema:
                    type: string
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/api.rulego.v1.DeleteRunLogsResponse'
...
components:
    schemas:
...
        api.rulego.v1.DebugLogEntry:
            type: object
            properties:
                chainId:
                    type: string
                nodeId:
                    type: string
                message:
                    type: string
                timestamp:
                    type: string
                level:
                    type: string
                metadata:
                    type: object
                    additionalProperties:
                        type: string
            description: 调试日志条目
...
        api.rulego.v1.DeleteRunLogsResponse:
            type: object
            properties:
                success:
                    type: boolean
                message:
                    type: string
                deletedCount:
                    type: integer
                    format: int32
            description: 删除运行日志响应
...
        api.rulego.v1.GetDebugLogsResponse:
            type: object
            properties:
                logs:
                    type: array
                    items:
                        $ref: '#/components/schemas/api.rulego.v1.DebugLogEntry'
                total:
                    type: integer
                    format: int32
                page:
                    type: integer
                    format: int32
                pageSize:
                    type: integer
                    format: int32
            description: 获取调试日志响应
...
        api.rulego.v1.GetLocalesResponse:
            type: object
            properties:
                locales:
                    type: object
                    additionalProperties:
                        type: string
            description: 获取本地化配置响应
...
        api.rulego.v1.ListRunLogsResponse:
            type: object
            properties:
                logs:
                    type: array
                    items:
                        $ref: '#/components/schemas/api.rulego.v1.RunLogEntry'
                total:
                    type: integer
                    format: int32
                page:
                    type: integer
                    format: int32
                pageSize:
                    type: integer
                    format: int32
            description: 获取运行日志列表响应
...
        api.rulego.v1.RunLogEntry:
            type: object
            properties:
                id:
                    type: string
                chainId:
                    type: string
                msgType:
                    type: string
                data:
                    type: string
                result:
                    type: string
                startTime:
                    type: string
                endTime:
                    type: string
                status:
                    type: string
                errorMessage:
                    type: string
                username:
                    type: string
            description: 运行日志条目
        api.rulego.v1.SaveLocalesRequest:
            type: object
            properties:
                lang:
                    type: string
                locales:
                    type: object
                    additionalProperties:
                        type: string
                username:
                    type: string
            description: 保存本地化配置请求
        api.rulego.v1.SaveLocalesResponse:
            type: object
            properties:
                success:
                    type: boolean
                message:
                    type: string
            description: 保存本地化配置响应
...